chore: rename animals controller to animal.
